requirements.txt
Flask
slackeventsapi
requests
beautifulsoup4
slackclient
python-dotenv